requirements.txt
Django==5.2.12
google-generativeai==0.8.3
python-dotenv==1.0.1
Pillow==10.3.0
PyMuPDF==1.24.4
google-genai==1.3.0
